Modernize based on latest approach to multiple nodes.

+++ examples/sensornet/nodeconfig.h
@@ -9,7 +9,7 @@
 #ifndef __NODECONFIG_H__
 #define __NODECONFIG_H__
 
-uint8_t nodeconfig_read(void);
+uint16_t nodeconfig_read(void);
 void nodeconfig_listen(void);
 
 #endif // __NODECONFIG_H__
